requirements.txt
pandas==2.0.3
numpy==1.24.3
langchain==0.1.14
langchain-community==0.0.30
chromadb==0.4.24
sentence-transformers==2.2.2
huggingface-hub==0.19.4
streamlit==1.28.1
python-dotenv==1.0.0
requests==2.31.0
groq==0.5.0